+++ src/app/(frontend)/volumes/[slug]/page.tsx
@@ -1,3 +1,6 @@
+import config from "@payload-config"
+import { cache } from "react"
+
 import { ArticleCard } from "@/components/ArticleCard"
 import { AuthorList } from "@/components/Authors/AuthorList"
 import { JsonLd } from "@/components/JsonLd"
@@ -6,15 +9,52 @@ import { LivePreviewListener } from "@/components/LivePreviewListener"
 import { PayloadRedirects } from "@/components/PayloadRedirects"
 import RichText from "@/components/RichText"
 import { Separator } from "@/components/ui/separator"
-import type { Article } from "@/payload-types"
-import { queryVolumeBySlug, queryVolumeSlugs } from "@/utilities/contentQueries"
+import type { Article, Volume } from "@/payload-types"
 import { formatDateTime } from "@/utilities/formatDateTime"
 import { generateMeta } from "@/utilities/generateMeta"
 import { buildBreadcrumbJsonLd, buildVolumeJsonLd } from "@/utilities/structuredData"
 import { toRoman } from "@/utilities/toRoman"
 import type { Metadata } from "next"
 import { draftMode } from "next/headers"
+import { getPayload } from "payload"
 import React from "react"
+
+const queryVolumeSlugs = cache(async (): Promise<{ slug: string | null | undefined }[]> => {
+  const payload = await getPayload({ config })
+  const { docs } = await payload.find({
+    collection: "volumes",
+    draft: false,
+    limit: 1000,
+    overrideAccess: false,
+    pagination: false,
+    select: {
+      slug: true,
+    },
+  })
+
+  return docs.map(({ slug }) => ({ slug }))
+})
+
+const queryVolumeBySlug = cache(async (slug: string): Promise<Volume | null> => {
+  const { isEnabled: draft } = await draftMode()
+  const payload = await getPayload({ config })
+
+  const { docs } = await payload.find({
+    collection: "volumes",
+    draft,
+    limit: 1,
+    overrideAccess: draft,
+    pagination: false,
+    where: {
+      slug: {
+        equals: slug,
+      },
+    },
+    depth: 2,
+  })
+
+  return docs[0] || null
+})
 
 export async function generateStaticParams(): Promise<{ slug: string | null | undefined }[]> {
   return queryVolumeSlugs()
