+++ src/collections/helpers.ts
@@ -1,17 +1,4 @@
-import {
-  MetaDescriptionField,
-  MetaImageField,
-  MetaTitleField,
-  OverviewField,
-  PreviewField,
-} from "@payloadcms/plugin-seo/fields"
-import type {
-  CollectionBeforeChangeHook,
-  CollectionConfig,
-  FieldHook,
-  Tab,
-  TypeWithID,
-} from "payload"
+import type { CollectionBeforeChangeHook, CollectionConfig, FieldHook, TypeWithID } from "payload"
 
 import { generatePreviewPath } from "@/utilities/generatePreviewPath"
 
@@ -41,32 +28,6 @@ export function previewAdminConfig(
         collection,
         req,
       }),
-  }
-}
-
-export function seoTab(): Tab {
-  return {
-    name: "meta",
-    label: "SEO",
-    fields: [
-      OverviewField({
-        titlePath: "meta.title",
-        descriptionPath: "meta.description",
-        imagePath: "meta.image",
-      }),
-      MetaTitleField({
-        hasGenerateFn: true,
-      }),
-      MetaImageField({
-        relationTo: "media",
-      }),
-      MetaDescriptionField({}),
-      PreviewField({
-        hasGenerateFn: true,
-        titlePath: "meta.title",
-        descriptionPath: "meta.description",
-      }),
-    ],
   }
 }
